Time: 1 ms (99.95%) | Memory: 77.2 MB (66.54%) - LeetSync

+++ 53-maximum-subarray/maximum-subarray.java
@@ -1,7 +1,7 @@
 class Solution {
     public int maxSubArray(int[] nums) {
-        long maxi=Long.MIN_VALUE;
-        long sum=0;
+        int maxi=Integer.MIN_VALUE;
+        int sum=0;
         for(int i=0;i<nums.length;i++){
             sum+=nums[i];
             if(sum>maxi){
@@ -11,7 +11,7 @@ class Solution {
                 sum=0;
             }
         }
-        return (int)maxi;
+        return maxi;
         
     }
 }
